scrape/requirements.txt
pandas
numpy
requests
spotipy
python-dotenv
bs4
sqlalchemy
psycopg2